feat: Switch RealAuthService to TestAuthService
- We don't need to confirm mail while testing.
